屏蔽了中文文件的上传和操作

+++ dir.func.php
@@ -32,19 +32,25 @@ function dirsize($path){
     return $sum;
 }
 
-function delFolder($path){
-    $handle=opendir($path);
-    while(($item=readdir($handle))!==false){
-        if($item!="."&&$item!=".."){
-            if(is_file($path."/".$item)){
-                unlink($path."/".$item);
-            }
-            if(is_dir($path."/".$item)){
-                delFolder($path."/".$item);
+function delFolder($path)
+{
+    if (! is_file($filename)) {
+        alert("Cant operate Chinese items! Please contact your server master!", "file.php");
+        exit();
+    } else {
+        $handle = opendir($path);
+        while (($item = readdir($handle)) !== false) {
+            if ($item != "." && $item != "..") {
+                if (is_file($path . "/" . $item)) {
+                    unlink($path . "/" . $item);
+                }
+                if (is_dir($path . "/" . $item)) {
+                    delFolder($path . "/" . $item);
+                }
             }
         }
+        closedir($handle);
+        rmdir($path);
     }
-    closedir($handle);
-    rmdir($path);
 }
 ?>
